# feat(server): improve checkAlbumAccess query performance